+++ pkg/graph/resolvers/waiver_assessment_survey.go
@@ -5,16 +5,118 @@ import (
 
 	"github.com/google/uuid"
 	"github.com/jmoiron/sqlx"
+	"github.com/samber/lo"
 	"go.uber.org/zap"
 
 	"github.com/cms-enterprise/mint-app/pkg/authentication"
 	"github.com/cms-enterprise/mint-app/pkg/email"
 	"github.com/cms-enterprise/mint-app/pkg/models"
+	"github.com/cms-enterprise/mint-app/pkg/notifications"
 	"github.com/cms-enterprise/mint-app/pkg/shared/oddmail"
 	"github.com/cms-enterprise/mint-app/pkg/sqlutils"
 	"github.com/cms-enterprise/mint-app/pkg/storage"
 	"github.com/cms-enterprise/mint-app/pkg/storage/loaders"
 )
+
+// SeedSuggestedWaivers creates a suggested_waiver row for every common_waiver, giving a
+// new model plan the full set of suggestions before any survey answers are recorded.
+func SeedSuggestedWaivers(
+	np sqlutils.NamedPreparer,
+	logger *zap.Logger,
+	modelPlanID uuid.UUID,
+	principal authentication.Principal,
+) error {
+	commonWaivers, err := storage.CommonWaiverGetAll(np, logger)
+	if err != nil {
+		return err
+	}
+
+	for _, cw := range commonWaivers {
+		sw := models.NewSuggestedWaiver(principal.Account().ID, modelPlanID, cw.ID)
+		if _, err := storage.SuggestedWaiverCreate(np, logger, sw); err != nil {
+			return err
+		}
+	}
+	return nil
+}
+
+// RecalculateSuggestedWaivers replaces all suggested_waiver rows for the model plan based on
+// the current survey answers. A common waiver is included when:
+//   - its survey_question_field is nil (no mapping yet → always suggest), or
+//   - the corresponding survey field is nil (unanswered → still suggest), or
+//   - the corresponding survey field is true (model needs this waiver type).
+//
+// When real waiver→question mappings are available, populate common_waiver.survey_question_field
+// via a migration — no Go code changes needed.
+func RecalculateSuggestedWaivers(
+	np sqlutils.NamedPreparer,
+	logger *zap.Logger,
+	survey *models.WaiverAssessmentSurvey,
+	principal authentication.Principal,
+) error {
+	if _, err := storage.SuggestedWaiverDeleteAllByModelPlanID(np, logger, survey.ModelPlanID); err != nil {
+		return err
+	}
+
+	commonWaivers, err := storage.CommonWaiverGetAll(np, logger)
+	if err != nil {
+		return err
+	}
+
+	for _, cw := range commonWaivers {
+		if !shouldSuggestWaiver(cw, survey) {
+			continue
+		}
+		sw := models.NewSuggestedWaiver(principal.Account().ID, survey.ModelPlanID, cw.ID)
+		if _, err := storage.SuggestedWaiverCreate(np, logger, sw); err != nil {
+			return err
+		}
+	}
+	return nil
+}
+
+// shouldSuggestWaiver returns false only when the waiver has a survey_question_field mapping
+// AND that survey question was explicitly answered false.
+func shouldSuggestWaiver(cw *models.CommonWaiver, survey *models.WaiverAssessmentSurvey) bool {
+	if cw.SurveyQuestionField == nil {
+		return true
+	}
+	val := surveyBoolField(survey, *cw.SurveyQuestionField)
+	if val == nil {
+		return true // unanswered — keep suggesting
+	}
+	return *val
+}
+
+// surveyBoolField looks up a boolean field on WaiverAssessmentSurvey by its DB column name.
+// Returns nil if the column name is unknown or the field is unset.
+// Add new cases here when new survey questions are introduced.
+func surveyBoolField(s *models.WaiverAssessmentSurvey, col string) *bool {
+	switch col {
+	case "modifies_medicare_savings_programs":
+		return s.ModifiesMedicareSavingsPrograms
+	case "bundles_payments":
+		return s.BundlesPayments
+	case "offers_risk_sharing_arrangements":
+		return s.OffersRiskSharingArrangements
+	case "impacts_site_of_care_payments":
+		return s.ImpactsSiteOfCarePayments
+	case "modifies_care_team_scope_of_practice":
+		return s.ModifiesCareTeamScopeOfPractice
+	case "modifies_care_delivery_with_claims_based_payments":
+		return s.ModifiesCareDeliveryWithClaimsBasedPayments
+	case "modifies_quality_measurements_or_payments_via_waivers":
+		return s.ModifiesQualityMeasurementsOrPaymentsViaWaivers
+	case "impacts_medicaid_only_beneficiaries":
+		return s.ImpactsMedicaidOnlyBeneficiaries
+	case "impacts_home_community_based_service_payments":
+		return s.ImpactsHomeCommunityBasedServicePayments
+	case "impacts_managed_care_waivers":
+		return s.ImpactsManagedCareWaivers
+	default:
+		return nil
+	}
+}
 
 // WaiverAssessmentSurveyGetByModelPlanID returns the waiver assessment survey associated with a model plan via dataloader
 func WaiverAssessmentSurveyGetByModelPlanID(ctx context.Context, modelPlanID uuid.UUID) (*models.WaiverAssessmentSurvey, error) {
@@ -118,4 +220,127 @@ func SuggestedWaiversGetByModelPlanID(ctx context.Context, modelPlanID uuid.UUID
 // CommonWaiverGetByID returns a common waiver by ID via dataloader
 func CommonWaiverGetByID(ctx context.Context, id uuid.UUID) (*models.CommonWaiver, error) {
 	return loaders.CommonWaiver.ByID.Load(ctx, id)
+}
+
+// TrySendWaiverAssessmentSurveyNotifications fires in-app and email notifications when a
+// waiver assessment survey transitions to COMPLETE. Errors are logged but do not fail the
+// parent transaction since notification delivery is best-effort.
+func TrySendWaiverAssessmentSurveyNotifications(
+	ctx context.Context,
+	survey *models.WaiverAssessmentSurvey,
+	logger *zap.Logger,
+	emailService oddmail.EmailService,
+	emailAddressBook email.AddressBook,
+	principal authentication.Principal,
+	np sqlutils.NamedPreparer,
+) {
+	modelPlan, err := ModelPlanGetByIDLOADER(ctx, survey.ModelPlanID)
+	if err != nil {
+		logger.Error("waiver assessment survey notification: failed to load model plan", zap.Error(err))
+		return
+	}
+
+	uans, err := storage.UserAccountGetNotificationPreferencesForWaiverAssessmentSurveyMarkedComplete(np, survey.ModelPlanID)
+	if err != nil {
+		logger.Error("waiver assessment survey notification: failed to get notification preferences", zap.Error(err))
+		return
+	}
+
+	emailUANs, inAppUANs := models.FilterNotificationPreferences(uans)
+
+	_, err = notifications.ActivityWaiverAssessmentSurveyMarkedCompleteCreate(
+		ctx,
+		principal.Account().ID,
+		np,
+		inAppUANs,
+		survey.ModelPlanID,
+		survey.ID,
+		principal.Account().ID,
+	)
+	if err != nil {
+		logger.Error("waiver assessment survey notification: failed to create activity", zap.Error(err))
+		return
+	}
+
+	go func() {
+		if emailService == nil {
+			return
+		}
+
+		// Send to MINT team (no footer)
+		if err := sendWaiverAssessmentSurveyMarkedCompleteEmail(
+			emailService,
+			emailAddressBook,
+			modelPlan,
+			emailAddressBook.MINTTeamEmail,
+			principal.Account().CommonName,
+			false,
+		); err != nil {
+			logger.Error("waiver assessment survey notification: failed to send MINT team email", zap.Error(err))
+			return
+		}
+
+		// Send to opted-in users (with footer)
+		if err := sendWaiverAssessmentSurveyMarkedCompleteEmails(
+			emailService,
+			emailAddressBook,
+			emailUANs,
+			modelPlan,
+			principal.Account().CommonName,
+			true,
+		); err != nil {
+			logger.Error("waiver assessment survey notification: failed to send user emails", zap.Error(err))
+		}
+	}()
+}
+
+func getWaiverAssessmentSurveyMarkedCompleteEmailContent(
+	emailService oddmail.EmailService,
+	modelPlan *models.ModelPlan,
+	markedCompleteByCommonName string,
+	showFooter bool,
+) (string, string, error) {
+	subjectContent := email.WaiverAssessmentSurveyCompletedSubjectContent{
+		ModelName: modelPlan.ModelName,
+	}
+	bodyContent := email.WaiverAssessmentSurveyCompletedBodyContent{
+		ClientAddress:                   emailService.GetConfig().GetClientAddress(),
+		ModelName:                       modelPlan.ModelName,
+		ModelID:                         modelPlan.GetModelPlanID().String(),
+		MarkedCompletedByUserCommonName: markedCompleteByCommonName,
+	}
+	return email.WaiverAssessmentSurvey.Completed.GetContent(subjectContent, bodyContent)
+}
+
+func sendWaiverAssessmentSurveyMarkedCompleteEmail(
+	emailService oddmail.EmailService,
+	addressBook email.AddressBook,
+	modelPlan *models.ModelPlan,
+	userEmail string,
+	markedCompleteByCommonName string,
+	showFooter bool,
+) error {
+	subject, body, err := getWaiverAssessmentSurveyMarkedCompleteEmailContent(emailService, modelPlan, markedCompleteByCommonName, showFooter)
+	if err != nil {
+		return err
+	}
+	return emailService.Send(addressBook.DefaultSender, []string{userEmail}, nil, subject, "text/html", body)
+}
+
+func sendWaiverAssessmentSurveyMarkedCompleteEmails(
+	emailService oddmail.EmailService,
+	addressBook email.AddressBook,
+	receivers []*models.UserAccountAndNotificationPreferences,
+	modelPlan *models.ModelPlan,
+	markedCompleteByCommonName string,
+	showFooter bool,
+) error {
+	subject, body, err := getWaiverAssessmentSurveyMarkedCompleteEmailContent(emailService, modelPlan, markedCompleteByCommonName, showFooter)
+	if err != nil {
+		return err
+	}
+	receiverEmails := lo.Map(receivers, func(pref *models.UserAccountAndNotificationPreferences, _ int) string {
+		return pref.Email
+	})
+	return emailService.Send(addressBook.DefaultSender, []string{}, nil, subject, "text/html", body, oddmail.WithBCC(receiverEmails))
 }
